public/solana_bots/secondary-wallet-tracking/requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
typing_extensions==4.12.2
yarl==1.9.4
